read category for preview page
